requirements.txt
elevenlabs
streamlit
Pillow
langchain